+++ src/game/traps/falling_block.h
@@ -7,6 +7,6 @@
 void SetupFallingBlock(OBJECT_INFO *obj);
 void FallingBlockControl(int16_t item_num);
 void FallingBlockFloor(
-    ITEM_INFO *item, int32_t x, int32_t y, int32_t z, int32_t *height);
+    ITEM_INFO *item, int32_t x, int32_t y, int32_t z, int16_t *height);
 void FallingBlockCeiling(
-    ITEM_INFO *item, int32_t x, int32_t y, int32_t z, int32_t *height);
+    ITEM_INFO *item, int32_t x, int32_t y, int32_t z, int16_t *height);
